client library - Use -fno-builtin to eliminate "conflicting type" warnings, and then eliminated the use of size_t in favor of "int", because size_t requires a dependency on external include files.
git-svn-id: svn://coreboot.org/openfirmware@2286 1552c027-8020-0410-b4b5-a757f869b4ce

+++ clients/lib/mem.c
@@ -3,14 +3,14 @@
 #include "1275.h"
 
 VOID
-memcpy(char *to, char *from, size_t len)
+memcpy(char *to, char *from, int len)
 {
 	while (len--)
 		*to++ = *from++;
 }
 
 VOID
-memset(char *cp, int c, size_t len)
+memset(char *cp, int c, int len)
 {
 	while (len--)
 		*(cp + len) = c;
